requirements.txt
Flask~=2.3.0
Flask-SQLAlchemy~=3.0.0
Flask-Migrate~=4.0.0
Flask-Login~=0.6.0
Flask-WTF~=1.1.0
WTForms~=3.0.0
python-dotenv~=1.0.0
slack-sdk~=3.23.0
python-dateutil~=2.8.0
pytz~=2023.3
gunicorn~=21.2.0
google-auth~=2.23.0
google-auth-oauthlib~=1.1.0
google-auth-httplib2~=0.1.0
requests~=2.31.0
